// frontend/src/interfaces/FormCreateAd/FromCreateAd.ts
export interface IFormCreateAd {
	typeAd: string;
	title: string;
	description: string;
	year: number;
	mileage: number;
	price: string;
	typeVehicle: string;
	urlCoverImage: string;
	urlImage1?: string;
	urlImage2?: string;
	urlImage3?: string;
	urlImage4?: string;
	urlImage5?: string;
	urlImage6?: string;
}